Add pre RA peephole phase to Z Codegen
Add phase for peephole optimizations before register
allocation to Z codegen.
