+++ compiler/z/CMakeLists.txt
@@ -36,6 +36,7 @@ compiler_library(z
 	${CMAKE_CURRENT_LIST_DIR}/codegen/S390GenerateInstructions.cpp
 	${CMAKE_CURRENT_LIST_DIR}/codegen/S390HelperCallSnippet.cpp
 	${CMAKE_CURRENT_LIST_DIR}/codegen/S390Instruction.cpp
+	${CMAKE_CURRENT_LIST_DIR}/codegen/S390Peephole.cpp
 	${CMAKE_CURRENT_LIST_DIR}/codegen/OMRLinkage.cpp
 	${CMAKE_CURRENT_LIST_DIR}/codegen/SystemLinkage.cpp
 	${CMAKE_CURRENT_LIST_DIR}/codegen/S390OutOfLineCodeSection.cpp
