+++ compiler/z/codegen/OMRCodeGenPhase.hpp
@@ -1,5 +1,5 @@
 /*******************************************************************************
- * Copyright (c) 2000, 2016 IBM Corp. and others
+ * Copyright (c) 2000, 2018 IBM Corp. and others
  *
  * This program and the accompanying materials are made available under
  * the terms of the Eclipse Public License 2.0 which accompanies this
@@ -51,12 +51,15 @@ class OMR_EXTENSIBLE CodeGenPhase : public OMR::CodeGenPhase
    public:
    static void performMarkLoadAsZeroOrSignExtensionPhase(TR::CodeGenerator * cg, TR::CodeGenPhase *);
    static void performSetBranchOnCountFlagPhase(TR::CodeGenerator * cg, TR::CodeGenPhase *);
+   static void performPreRAPeepholePhase(TR::CodeGenerator * cg, TR::CodeGenPhase *);
 
    // override base class implementation because new phases are being added
    static int getNumPhases();
    const char * getName();
    static const char* getName(PhaseValue phase);
 
+   // override base class to call new post RA peephole
+   static void performPeepholePhase(TR::CodeGenerator * cg, TR::CodeGenPhase *);
    };
 }
