+++ compiler/z/codegen/OMRCodeGenPhaseEnum.hpp
@@ -1,5 +1,5 @@
 /*******************************************************************************
- * Copyright (c) 2000, 2016 IBM Corp. and others
+ * Copyright (c) 2000, 2018 IBM Corp. and others
  *
  * This program and the accompanying materials are made available under
  * the terms of the Eclipse Public License 2.0 which accompanies this
@@ -30,4 +30,5 @@
 
 markLoadAsZeroOrSignExtension,
 SetBranchOnCountFlagPhase,
-LastOMRZPhase = SetBranchOnCountFlagPhase,
+PreRAPeepholePhase,
+LastOMRZPhase = PreRAPeepholePhase,
